pyproject.toml
[project]
name = "notebook-17-ia-generativa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "icrawler>=0.6.10",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
